Allowing java_wire_library to specify sources that don't exist.
This enables us to pull sources in from resolved jar files. This is
hacky for a couple reasons: (1) It causes mispelled sources to fail
last rather than fail fast, because they are assumed to be virtual
until the wire compiler actually crashes. (2) Compiling sources
from jar files is itself a weird and hacky thing to do.

The most common example of protos we pull in from jars is the
google descriptor.proto, which we pull in from:

  3rdparty:com.google.protobuf.protobuf-java

Because this is so hacky, I'm not currently intending to upstream
these changes.

I've tried alternate techniques of getting this to work, including
having an intermediate java_wire_library() set its sources to
from_target, with the actual java_wire_library target depending on
that one, but for whatever reason the compilation fails unless the
protos are compiled all-together. I'm not sure why, but it may have
something to do with the service factories we use internally.