Adjust the output file locations for the Antlr task.
The Antlr rules don't currently lead to a happy IntelliJ configuration. This PR is one way to fix it, although I'm not really sure what the best approach is.

The root problem is that Antlr dumps its output files into a path determined by the relative path specified for the input file. This PR takes the simple approach of moving all the files once they have been generated.

As a possible alternative, the files could be left in their current location, and the IntelliJ support could be modified to generate a better source root for the resulting code. It seemed slightly better, though, to try and fix the file locations earlier in the pipeline, right when they are generated. Antlr has predictable output, so it's not too hard to just fix it up.

Testing Done:
I checked the behavior with the checked-in Antlr samples:
```
./pants  idea ./examples/src/java/org/pantsbuild/example/antlr3::
./pants  idea ./examples/src/java/org/pantsbuild/example/antlr4::
```
I also ran the the `unit-test.sh` script.

Travis CI has passed: https://travis-ci.org/pantsbuild/pants/builds/152249047

Bugs closed: 3787

Reviewed at https://rbcommons.com/s/twitter/r/4161/